Refactor time reporting and validation
- Remove obsolete helper modules (item_data_helper, time_report_helper, time_update_helper, time_validation_helper).
- Integrate validation logic into ValidationEditor via new TimeValidationHelpers module.
- Add export_csv and export_icalendar methods to TimeReport.
- Update .reek.yml with exclusions for TooManyMethods, LongParameterList, and DataClump.
- Adjust Table to drop TimeReportHelper include.
- Update specs to reflect new class responsibilities.
